[app] A few book cover rendering tweaks
Edge filling is only performed when both edges consist predominantly
of the same color and the aspect ratio significantly differs from
the target one. Otherwise the cover is stretched pretty much as before
except that stretched image is now centered. That seems to works nicely
in all the cases which I've encoutered so far.